Fixed error with theme generator
Theme page was adding blank/missing images when you hover over them.
This was do to a bad check to see if a varriable($sshot_2, $sshot_3) were empty.
They were both being checked against $sshot_1 then being filled with a number "2-" and "3-" respectivley.
This is what was causing the blank image to show on hover.

Change-Id: I6c9f62f1ff73222f3adb7d317f2a397d3f84528a
Reviewed-on: http://gerrit.rockbox.org/309
Reviewed-by: Michael Rodger <m.rodger@sae.edu>
Reviewed-by: Torne Wuff <torne@wolfpuppy.org.uk>