Created redirect to most recent update of theme
If you visit a theme that has been updated
you will be redirected to the newest verion of the theme.
This will be helpful if other websites link to a theme
and it is updated. Or if someone links from the forum.
Then you wont have to update your link in the forum every
time you update your theme.

Change-Id: I7737d0cb2b191d5397bf207aad0d6d17003d2025
Reviewed-on: http://gerrit.rockbox.org/343
Reviewed-by: Torne Wuff <torne@wolfpuppy.org.uk>